skills_finder_web/requirements.txt
boto3==1.20.0
Django==3.2.9
pandas==1.3.4
psycopg2-binary==2.9.2
python-dotenv==0.19.2
smart-open==5.2.1
